+++ src/main/java/sopt/org/homepage/project/ProjectService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import sopt.org.homepage.common.dto.PaginateResponseDto;
 import sopt.org.homepage.common.util.ArrayUtil;
 import sopt.org.homepage.internal.playground.PlaygroundService;
-import sopt.org.homepage.project.dto.GetProjectsRequestDto;
-import sopt.org.homepage.project.dto.ProjectDetailResponseDto;
-import sopt.org.homepage.project.dto.ProjectsResponseDto;
+import sopt.org.homepage.project.dto.request.GetProjectsRequestDto;
+import sopt.org.homepage.project.dto.response.ProjectDetailResponseDto;
+import sopt.org.homepage.project.dto.response.ProjectsResponseDto;
 
 @RequiredArgsConstructor
 @Service
