+++ src/main/java/sopt/org/homepage/project/dto/record/Link.java
@@ -1,4 +1,4 @@
-package sopt.org.homepage.project.dto;
+package sopt.org.homepage.project.dto.record;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import sopt.org.homepage.project.dto.type.LinkType;
