+++ src/main/java/sopt/org/homepage/project/dto/request/GetProjectsRequestDto.java
@@ -1,4 +1,4 @@
-package sopt.org.homepage.project.dto;
+package sopt.org.homepage.project.dto.request;
 
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.Getter;
